requirements.txt
aiogram==2.25.1
requests>=2.31,<3
python-dotenv>=1.0,<2
